Reworking for cvenv
