+++ ons_ras_common/ons_version.py
@@ -1,1 +1,1 @@
-__version__ = '0.2.19'
+__version__ = '0.2.20'
